+++ src/gridcv.jl
@@ -71,9 +71,8 @@ res = gridcvlv(Xtrain, ytrain; segm = segm,
     score = rmsep, fun = plskern, nlv = nlv).res
 u = findall(res.y1 .== minimum(res.y1))[1] 
 res[u, :]
-
-lines(res.nlv, res.y1,
-    axis = (xlabel = "Nb. LVs", ylabel = "RMSEP"))
+plotscore(res.nlv, res.y1;
+    xlabel = "Nb. LVs", ylabel = "RMSEP").f
 
 fm = plskern(Xtrain, ytrain; nlv = res.nlv[u]) ;
 pred = Jchemo.predict(fm, Xtest).pred 
@@ -93,9 +92,9 @@ res = gridcvlv(Xtrain, ytrain; segm = segm,
     score = rmsep, fun = lwplsr, pars = pars, nlv = nlv, verbose = true).res
 u = findall(res.y1 .== minimum(res.y1))[1] 
 res[u, :]
-
-lines(res.nlv, res.y1,
-    axis = (xlabel = "Nb. LVs", ylabel = "RMSEP"))
+group = string.("h=", res.h, " k=", res.k)
+plotscore(res.nlv, res.y1, group;
+    xlabel = "Nb. LVs", ylabel = "RMSECV").f
 
 fm = lwplsr(Xtrain, ytrain;
     nlvdis = res.nlvdis[u], metric = res.metric[u],
@@ -113,9 +112,8 @@ res = gridcvlb(Xtrain, ytrain; segm = segm,
     score = rmsep, fun = rr, lb = lb).res
 u = findall(res.y1 .== minimum(res.y1))[1] 
 res[u, :]
-
-lines(log.(res.lb), res.y1,
-    axis = (xlabel = "Nb. LVs", ylabel = "RMSEP"))
+plotscore(log.(res.lb), res.y1;
+    xlabel = "Lambda", ylabel = "RMSECV").f
 
 fm = rr(Xtrain, ytrain; lb = res.lb[u]) ;
 pred = Jchemo.predict(fm, Xtest).pred 
@@ -134,9 +132,9 @@ res = gridcvlb(Xtrain, ytrain; segm = segm,
     score = rmsep, fun = krr, pars = pars, lb = lb).res
 u = findall(res.y1 .== minimum(res.y1))[1] 
 res[u, :]
-
-lines(log.(res.lb), res.y1,
-    axis = (xlabel = "Nb. LVs", ylabel = "RMSEP"))
+group = string.("gamma=", res.gamma)
+plotscore(log.(res.lb), res.y1, group;
+    xlabel = "Lambda", ylabel = "RMSECV").f
 
 fm = krr(Xtrain, ytrain; gamma = res.gamma[u], lb = res.lb[u]) ;
 pred = Jchemo.predict(fm, Xtest).pred