+++ src/gridscore.jl
@@ -72,9 +72,8 @@ res = gridscorelv(Xcal, ycal, Xval, yval;
     score = rmsep, fun = plskern, nlv = nlv)
 u = findall(res.y1 .== minimum(res.y1))[1] 
 res[u, :]
-
-lines(res.nlv, res.y1,
-    axis = (xlabel = "Nb. LVs", ylabel = "RMSEP"))
+plotscore(res.nlv, res.y1;
+    xlabel = "Nb. LVs", ylabel = "RMSEP").f
 
 fm = plskern(Xtrain, ytrain; nlv = res.nlv[u]) ;
 pred = Jchemo.predict(fm, Xtest).pred 
@@ -91,9 +90,9 @@ res = gridscorelv(Xcal, ycal, Xval, yval;
     score = rmsep, fun = lwplsr, pars = pars, nlv = nlv, verbose = true)
 u = findall(res.y1 .== minimum(res.y1))[1] 
 res[u, :]
-
-lines(res.nlv, res.y1,
-    axis = (xlabel = "Nb. LVs", ylabel = "RMSEP"))
+group = string.("h=", res.h, " k=", res.k)
+plotscore(res.nlv, res.y1, group;
+    xlabel = "Nb. LVs", ylabel = "RMSECV").f
 
 fm = lwplsr(Xtrain, ytrain;
     nlvdis = res.nlvdis[u], metric = res.metric[u],
@@ -108,9 +107,8 @@ res = gridscorelb(Xcal, ycal, Xval, yval;
     score = rmsep, fun = rr, lb = lb)
 u = findall(res.y1 .== minimum(res.y1))[1] 
 res[u, :]
-
-lines(log.(res.lb), res.y1,
-    axis = (xlabel = "Nb. LVs", ylabel = "RMSEP"))
+plotscore(log.(res.lb), res.y1;
+    xlabel = "Lambda", ylabel = "RMSECV").f
 
 fm = rr(Xtrain, ytrain; lb = res.lb[u]) ;
 pred = Jchemo.predict(fm, Xtest).pred 
@@ -126,9 +124,9 @@ res = gridscorelb(Xcal, ycal, Xval, yval;
     score = rmsep, fun = krr, pars = pars, lb = lb)
 u = findall(res.y1 .== minimum(res.y1))[1] 
 res[u, :]
-
-lines(log.(res.lb), res.y1,
-    axis = (xlabel = "Nb. LVs", ylabel = "RMSEP"))
+group = string.("gamma=", res.gamma)
+plotscore(log.(res.lb), res.y1, group;
+    xlabel = "Lambda", ylabel = "RMSECV").f
 
 fm = krr(Xtrain, ytrain; gamma = res.gamma[u], lb = res.lb[u]) ;
 pred = Jchemo.predict(fm, Xtest).pred